feat: split code on modules
