+++ package.json
@@ -13,9 +13,13 @@
     "keywords": [],
     "license": "ISC",
     "devDependencies": {
-        "nodemon": "^2.0.16"
+        "nodemon": "^2.0.16",
+        "typescript": "^5.3.3"
     },
     "dependencies": {
-        "ws": "^8.8.0"
+        "ws": "^8.8.0",
+        "ts-node": "^10.9.2",
+        "ts-node-dev": "^2.0.0",
+        "dotenv": "^16.4.1"
     }
 }
